requirements.txt
django
django-bootstrap4
uwsgi
djangorestframework
django-rest-registration
drf-extensions
django-filter
python-dateutil
